Update Sum Of Two String (A+B).cpp

+++ Algorithms-and-Common-Problems/Sum Of Two String (A+B).cpp
@@ -1,20 +1,20 @@
 string findSum(string str1, string str2){
     if (str1.length() > str2.length()) swap(str1, str2);
     string str;
-    int n1 = str1.length(), n2 = str2.length();
+    ll n1 = str1.length(), n2 = str2.length();
 
     reverse(str1.begin(), str1.end());
     reverse(str2.begin(), str2.end());
 
-    int carry = 0;
-    for (int i = 0; i < n1; i++) {
-        int sum = ((str1[i] - '0') + (str2[i] - '0') + carry);
+    ll carry = 0;
+    for (ll i = 0; i < n1; i++) {
+        ll sum = ((str1[i] - '0') + (str2[i] - '0') + carry);
         str.push_back(sum % 10 + '0');
         carry = sum / 10;
     }
 
-    for (int i = n1; i < n2; i++){
-        int sum = ((str2[i] - '0') + carry);
+    for (ll i = n1; i < n2; i++){
+        ll sum = ((str2[i] - '0') + carry);
         str += (sum % 10 + '0');
         carry = sum / 10;
     }
